+++ sound-mind-studio/tests/test_mind_waves_panel.h
@@ -20,4 +20,5 @@ private slots:
     void setMindWavesPreservesSelectionAndRedisplaysFromTheNewData();
     void setMindWavesDropsSelectionWhenTheIdIsGone();
     void clearSelectionDisablesBothEditorsAndEmits();
+    void contentIsInAResizableScrollAreaSoThePanelCanShrinkBelowItsFullHeight();
 };